// src/types/page.ts
export type Page =
  | 'Dashboard'
  | 'ลูกค้า'
  | 'ใบประเมิน'
  | 'ภาคสนาม'
  | 'ใบเสนอราคา'
  | 'ใบแจ้งหนี้'
  | 'ใบกำกับภาษี/ใบเสร็จรับเงิน'
  | 'ฟอร์ม'
  | 'หมวดหมู่'
  | 'สินค้า/บริการ'
  | 'แพ็กเกจ'
  | 'คลังสินค้า'
  | 'ผู้จัดจำหน่าย'
  | 'รับเข้า'
  | 'เบิกสินค้า/อุปกรณ์ และค่าใช้จ่าย'
  | 'โอนย้าย'
  | 'ปรับปรุง Stock'
  | 'คืนสินค้า'
  | 'ผู้ใช้งาน'
  | 'จัดการบทบาท'
  | 'การแจ้งเตือน'
  | 'รายงาน'
  | 'เบิกสินค้าคืนผู้จำหน่าย'
  | 'รายงานรายได้ (รายเดือน)'
  | 'รายได้ออกใบกำกับ(รายเดือน)'
  | 'ค่าใช้จ่ายทางอ้อม'
  | 'บัญชีเงินสดรายวัน'
  | 'ค่าใช้จ่ายทางตรง'
  | 'ยอดขาย(รายเดือน)'
  | 'สรุปยอดขาย(รายเดือน)';